Make images clickable  + remove subscription ID info from sample image
Hi,

The inline images are very small and cannot be read. I added the option for the user to click on the images to go to the original large PNG screenshots so he or she can properly read the text of the screenshot.

Another issue that I'd appreciate if you could fix is the removal/truncation of the subscription ID info from the network-alert.png file. This subscription ID should not be publicly visible. 

Thanks,
Yoav